+++ src/muos/steps/select_disk.py
@@ -1,12 +1,10 @@
 from ..environment import Environment
 from ..get_disks import get_disks
-from ..step import Step
+from ..select_step import SelectStep
 
-class SelectDisk(Step):
+class SelectDisk(SelectStep):
     name: str = 'Select a disk'
 
     def main(self, environment: Environment) -> None:
-        super().main(environment)
-        choices = get_disks()
-        result = environment.fzf.prompt(choices, '--height=10 --layout=reverse')
-        environment.disk = result[0]
+        super().main(environment, get_disks())
+        environment.disk = self.choices[0]