+++ src/muos/steps/select_time_zone.py
@@ -1,14 +1,11 @@
 from ..environment import Environment
 from ..get_lines_as_list import get_lines_as_list
 from ..get_static_dir import get_static_dir
-from ..step import Step
+from ..select_step import SelectStep
 
-class SelectTimeZone(Step):
+class SelectTimeZone(SelectStep):
     name: str = 'Select a time zone'
 
     def main(self, environment: Environment) -> None:
-        super().main(environment)
-        timezones_txt = get_static_dir() / 'timezones.txt'
-        choices = get_lines_as_list(timezones_txt)
-        result = environment.fzf.prompt(choices, '--height=10 --layout=reverse')
-        environment.time_zone = result[0]
+        super().main(environment, get_lines_as_list(get_static_dir() / 'timezones.txt'))
+        environment.time_zone = self.choices[0]